# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FactSnap API",
        "description": "API for managing educational categories, packages, and facts.\n *\n * # Authentication\n * \n * This API supports **dual authentication**:\n * \n * ## Firebase Authentication (for regular users)\n * - **Token Type**: Firebase ID Token\n * - **Format**: `Authorization: Bearer <firebase_id_token>`\n * - **Access**: 'free', 'loggedIn', and 'premium' packages (if premium user)\n * \n * ## Sanctum Authentication (for guest users)  \n * - **Token Type**: Sanctum Bearer Token (10-year expiration)\n * - **Format**: `Authorization: Bearer <sanctum_token>`\n * - **Access**: 'free' and 'loggedIn' packages ('premium' if created as premium guest)\n * \n * ## Package Access Levels\n * - **free**: No authentication required\n * - **loggedIn**: Requires authentication (Firebase OR Sanctum)\n * - **premium**: Requires premium user status",
        "contact": {
            "email": "support@factsnap.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Factsnap API Server"
        }
    ],
    "paths": {
        "/api/auth/guest": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Create a guest user",
                "description": "Create a new guest user with a long-lived Sanctum token (10 years expiration). Guest users can subscribe to packages based on their access level.",
                "operationId": "createGuest",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "is_premium": {
                                        "description": "Whether this guest should have premium access",
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Guest user created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Guest user created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "description": "Sanctum authentication token",
                                                    "type": "string"
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "format": "uuid"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Guest ABC123"
                                                        },
                                                        "is_guest": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "is_premium": {
                                                            "type": "boolean",
                                                            "example": false
                                                        },
                                                        "abilities": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "enum": [
                                                                    "guest",
                                                                    "premium"
                                                                ]
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "token_expires_at": {
                                                    "description": "Token expiration date (10 years from now)",
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "firebaseAuth": {
                "type": "http",
                "description": "Firebase ID Token obtained from Firebase SDK. Use in Authorization header: Bearer <firebase_id_token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctumAuth": {
                "type": "http",
                "description": "Sanctum Bearer Token for guest users. Get token from POST /api/auth/guest. Use in Authorization header: Bearer <sanctum_token>",
                "bearerFormat": "Token",
                "scheme": "bearer"
            },
            "dualAuth": {
                "type": "http",
                "description": "Accepts either Firebase ID tokens OR Sanctum tokens. Firebase is tried first, then Sanctum as fallback.",
                "scheme": "bearer"
            }
        }
    },
...
}